Create List Methods Module and Refactor Lists
