+++ app.rb
@@ -3,9 +3,11 @@ require_relative './teacher'
 require_relative './rental'
 require_relative './book'
 require_relative './create_methods'
+require_relative './list_methods'
 
 class App
   include CreateMethods
+  include ListMethods
 
   def initialize
     @people = []
@@ -57,29 +59,6 @@ class App
       list_all_people
     when 6
       list_person_rentals
-    end
-  end
-
-  def list_all_people
-    @people.each_with_index do |person, index|
-      puts "#{index}) [#{person.class.name}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
-    end
-  end
-
-  def list_all_books
-    @books.each_with_index { |book, index| puts %{#{index}) Title: \"#{book.title}\", Author: #{book.author}} }
-  end
-
-  def list_person_rentals
-    print 'ID of person: '
-    id = gets.chomp.to_i
-    puts 'Rentals: '
-    @people.each do |person|
-      next unless person.id == id
-
-      person.rentals.each do |rental|
-        puts %(Date: #{rental.date}, Book \"#{rental.book.title}\" by #{rental.book.author})
-      end
     end
   end
 end
